requirements.txt
pyaes==1.6.1
pyperclip==1.9.0
pyrotgfork==2.1.35
PySocks==1.7.1
PyTgCrypto==1.2.6
python-dotenv==1.0.1
TgCrypto==1.2.5